+++ lib/properties_intf.ml
@@ -2460,6 +2460,7 @@ and animation_timeline =
   | Name of string
   | Scroll of string
   | View of string
+  | Timelines of animation_timeline list
   | Initial
   | Inherit
   | Unset